[docs] Move modals & popovers to utils
