+++ docs/src/pages/utils/modals/SimpleModal.js
