+++ docs/src/pages/utils/modals/modals.md
@@ -8,7 +8,7 @@ The modal component provides a solid foundation for creating dialogs,
 popovers, lightboxes, or whatever else.
 The component renders its `children` node in front of a backdrop component.
 
-The `Modal` offers a few helpful features over using just a [`Portal`](/utils/portal)
+The `Modal` offers a few helpful features over using just a [`Portal`](/demos/portal)
 component and some styles:
 - Manages dialog stacking when one-at-a-time just isn't enough.
 - Creates a backdrop, for disabling interaction below the modal.
@@ -21,4 +21,4 @@ This component shares many concepts with [react-overlays](https://react-bootstra
 
 ## Simple modal
 
-{{"demo": "pages/demos/modals/SimpleModal.js"}}
+{{"demo": "pages/utils/modals/SimpleModal.js"}}
