+++ docs/src/pages/utils/popovers/popovers.md
@@ -14,10 +14,10 @@ When it is `anchorPosition`, the component will, instead of `anchorEl`,
 refer to the `anchorPosition` prop which you can adjust to set
 the position of the popover.
 
-{{"demo": "pages/demos/popovers/AnchorPlayground.js"}}
+{{"demo": "pages/utils/popovers/AnchorPlayground.js"}}
 
 ## Mouse over interaction
 
 We demonstrate how to use the `Popover` component as well as the [react-popper](https://github.com/souporserious/react-popper) package to implement a popover behavior based on the mouse over event.
 
-{{"demo": "pages/demos/popovers/MouseOverPopover.js"}}
+{{"demo": "pages/utils/popovers/MouseOverPopover.js"}}
