+++ docs/src/pages/utils/transitions/SimpleCollapse.js
