+++ docs/src/pages/utils/transitions/SimpleFade.js
