+++ docs/src/pages/utils/transitions/SimpleGrow.js
