+++ docs/src/pages/utils/transitions/SimpleSlide.js
