+++ docs/src/pages/utils/transitions/SimpleZoom.js
