+++ pages/utils/modals.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import withRoot from 'docs/src/modules/components/withRoot';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from 'docs/src/pages/demos/modals/modals.md';
+import markdown from 'docs/src/pages/utils/modals/modals.md';
 
 function Page() {
   return (
     <MarkdownDocs
       markdown={markdown}
       demos={{
-        'pages/demos/modals/SimpleModal.js': {
-          js: require('docs/src/pages/demos/modals/SimpleModal').default,
+        'pages/utils/modals/SimpleModal.js': {
+          js: require('docs/src/pages/utils/modals/SimpleModal').default,
           raw: preval`
 module.exports = require('fs')
-  .readFileSync(require.resolve('docs/src/pages/demos/modals/SimpleModal'), 'utf8')
+  .readFileSync(require.resolve('docs/src/pages/utils/modals/SimpleModal'), 'utf8')
 `,
         },
       }}
